This issue was reported by @anduriel. Both join buttons were shown in case of open-groups.

+++ src/resources/views/partials/join-button.blade.php
@@ -1,28 +1,28 @@
-@switch($seatgroup->type)
-  @case('open')
-    @switch($seatgroup->isMember(auth()->user()->group))
-      @case(false)
-      <form action="{{route('seatgroups.user.join', $seatgroup->id)}}" method="POST">
-        {{csrf_field()}}
-        <input name="_method3" type="hidden" value="PATCH">
-        <button type="submit" class="btn btn-sm btn-success pull-right" style="display: inline">{{trans('seatgroups::seat.seat_join_opengroup')}}</button>
-      </form>
-      @break
-      @case(true)
-      <form action="{{route('seatgroups.user.leave', $seatgroup->id)}}" method="DELETE">
-        {{csrf_field()}}
-        <button type="submit" class="btn btn-sm btn-danger pull-right" style="display: inline">{{trans('seatgroups::seat.seat_leave_opengroup')}}</button>
-      </form>
-      @break
-    @endswitch
-  @case('managed')
-    @if($seatgroup->onWaitlist())
-      <form action="{{route('seatgroups.user.leave', $seatgroup->id)}}" method="DELETE">
-        {{csrf_field()}}
-        <button type="submit" class="btn btn-sm btn-info pull-right" style="display: inline">{{trans('seatgroups::seat.seat_leave_waitlist')}}</button>
-      </form>
-      @break
-    @endif
+@if($seatgroup->type === 'open')
+  @switch($seatgroup->isMember(auth()->user()->group))
+    @case(false)
+    <form action="{{route('seatgroups.user.join', $seatgroup->id)}}" method="POST">
+      {{csrf_field()}}
+      <input name="_method3" type="hidden" value="PATCH">
+      <button type="submit" class="btn btn-sm btn-success pull-right" style="display: inline">{{trans('seatgroups::seat.seat_join_opengroup')}}</button>
+    </form>
+    @break
+    @case(true)
+    <form action="{{route('seatgroups.user.leave', $seatgroup->id)}}" method="DELETE">
+      {{csrf_field()}}
+      <button type="submit" class="btn btn-sm btn-danger pull-right" style="display: inline">{{trans('seatgroups::seat.seat_leave_opengroup')}}</button>
+    </form>
+    @break
+  @endswitch
+@endif
+
+@if($seatgroup->type === 'managed')
+  @if($seatgroup->onWaitlist())
+    <form action="{{route('seatgroups.user.leave', $seatgroup->id)}}" method="DELETE">
+      {{csrf_field()}}
+      <button type="submit" class="btn btn-sm btn-info pull-right" style="display: inline">{{trans('seatgroups::seat.seat_leave_waitlist')}}</button>
+    </form>
+  @else
     @switch($seatgroup->isMember(auth()->user()->group))
       @case(false)
       <form action="{{route('seatgroups.user.join', $seatgroup->id)}}" method="POST">
@@ -34,11 +34,12 @@
       @case(true)
       <form action="{{route('seatgroups.user.leave', $seatgroup->id)}}" method="DELETE">
         {{csrf_field()}}
-        <button type="submit" class="btn btn-sm btn-danger pull-right" style="display: inline">{{trans('seatgroups::seat.seat_leave_opengroup')}}</button>
+        <button type="submit" class="btn btn-sm btn-danger pull-right" style="display: inline">{{trans('seatgroups::seat.seat_leave_managedgroup')}}</button>
       </form>
       @break
     @endswitch
-@endswitch
+  @endif
+@endif
